Extract public key into utils
