+++ utils/Cargo.toml
@@ -8,16 +8,20 @@ edition = "2021"
 [dependencies]
 anyhow = "1"
 async-trait = "0.1"
+base58 = "0.2"
+blake2 = "0.10.5"
 bytesize = "1"
 bytesize-serde = "0.2"
 derivative = "2.2.0"
-parity-scale-codec = "3"
 derive_more = "0.99"
 futures = "0.3"
 jsonrpsee-core = "0.16"
-serde = { version = "1", features = ["derive"] }
-tokio = { version = "1", features = ["rt", "tracing"] }
 libp2p-core = "0.39"
+parity-scale-codec = "3"
+serde = { version = "1", features = ["derive"] }
+ss58-registry = "1.33"
+thiserror = "1"
+tokio = { version = "1", features = ["rt", "tracing"] }
 
 frame-support = { version = "4.0.0-dev", git = "https://github.com/subspace/substrate", rev = "9cf78129a2638d3f370868863d16f4fe32b4ad30" }
 frame-system = { version = "4.0.0-dev", git = "https://github.com/subspace/substrate", rev = "9cf78129a2638d3f370868863d16f4fe32b4ad30" }
